TextStylePropertyComponent.setIdPrefix pulled TextStyleLengthPropertyComponentLike
