+++ src/main/java/walkingkooka/spreadsheet/dominokit/value/textstyle/TextStylePropertyComponent.java
@@ -19,6 +19,7 @@ package walkingkooka.spreadsheet.dominokit.value.textstyle;
 
 import elemental2.dom.HTMLElement;
 import walkingkooka.naming.HasName;
+import walkingkooka.spreadsheet.dominokit.SpreadsheetElementIds;
 import walkingkooka.spreadsheet.dominokit.value.FormValueComponent;
 import walkingkooka.spreadsheet.dominokit.value.ValueWatcher;
 import walkingkooka.text.CaseKind;
@@ -35,6 +36,18 @@ import java.util.Optional;
 public interface TextStylePropertyComponent<E extends HTMLElement, V, C extends TextStylePropertyComponent<E, V, C>>
     extends FormValueComponent<E, V, C>,
     HasName<TextStylePropertyName<V>> {
+
+    default C setIdPrefix(final String idPrefix) {
+        return this.setId(
+            idPrefix +
+                CaseKind.KEBAB.change(
+                    this.name()
+                        .text(),
+                    CaseKind.CAMEL
+                ) +
+                SpreadsheetElementIds.TEXT_BOX
+        );
+    }
 
     default C setLabelFromPropertyName() {
         return this.setLabel(
